Adding support for global extensions
It seems that for some reason Google uses both gd* and g* prefixes to docs, slides, sheets and link files. Might be a localization thing, a regional thing, or simply a change at some point in time but the truth is that if you look it up there's both, though more commonly the g* prefix it seems. 

Adding support for that prefix. Hope it helps and that it works. 

Let me know if a new version can be published, assuming this is approved and doesn't break anything.