Update element identification for NC25
On NC25 (at least) the previous selector of #filesList was no longer available. The equivalent is now .files-fileList .

Unsure how this can be added so that it supports different NC versions - does it get published as a different version in the store?

Let me know if there's anything that should be changed here.